project/microservices/statistics/requirements.txt
faust-streaming
faust-streaming-rocksdb
